feature/rtrail_033 remove registration controller methods from user controller and create registration controller, change paths for login and registration pages, create Registration controller tests
