+++ rest/src/main/java/io/khasang/rtrail/controller/UserController.java
@@ -14,7 +14,6 @@ import java.util.List;
 
 /**
  * controller for REST CRUD services
- * provide registration and login page for validation
  *
  * @author Ilya Bogachev
  * @since 22.06.2018
@@ -24,9 +23,6 @@ import java.util.List;
 public class UserController {
     @Autowired
     private UserService userService;
-
-    @Autowired
-    private UserValidator userValidator;
 
     @RequestMapping(value = "/add", method = RequestMethod.POST, produces = "application/json;charset=utf-8")
     @ResponseBody
@@ -62,38 +58,5 @@ public class UserController {
     @ResponseBody
     public UserDTO deleteUser(@RequestParam(value = "id") String id) {
         return userService.deleteUser(Long.parseLong(id));
-    }
-
-    @RequestMapping(value = "/registration", method = RequestMethod.GET)
-    public String showRegistrationForm(Model model){
-        User userFromFormRegistration = new User();
-        model.addAttribute("user", userFromFormRegistration);
-        return "registration";
-    }
-
-    @RequestMapping(value = "/registration", method = RequestMethod.POST)
-    public String registration(@ModelAttribute("user") User userFromFormRegistration, BindingResult bindingResult, Model model){
-        userValidator.validate(userFromFormRegistration, bindingResult);
-
-        if (bindingResult.hasErrors()) {
-            return "registration";
-        }
-
-        userService.addUser(userFromFormRegistration);
-
-        return "redirect:/helloPage";
-    }
-
-    @RequestMapping(value = "/login", method = RequestMethod.GET)
-    public String login(Model model, String error, String logout) {
-        if (error != null) {
-            model.addAttribute("error", "Your email and password is invalid");
-        }
-
-        if (logout != null) {
-            model.addAttribute("message", "You have been logged out successfully");
-        }
-
-        return "login";
     }
 }
